ml/4-process_pdfs/requirements.txt
azure-ai-formrecognizer
azure.ai.documentintelligence
azure-core
azure-identity
azure-storage-blob
python-dotenv
markdownify
pypdf